Optimize BWC Model Sharing Resources
