+++ qa/restart-upgrade/src/test/java/org/opensearch/neuralsearch/bwc/restart/test/SemanticFieldIT.java
@@ -2,7 +2,7 @@
  * Copyright OpenSearch Contributors
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.opensearch.neuralsearch.bwc.restart;
+package org.opensearch.neuralsearch.bwc.restart.test;
 
 import java.util.Collections;
 import java.util.List;
